feat : 매칭 정보 저장을 위한 Pairs 구현
