+++ src/main/java/pairmatching/domain/pair/Pairs.java
@@ -1,0 +1,8 @@
+package pairmatching.domain.pair;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class Pairs {
+    private final Map<PairKey, Pair> pairs = new HashMap<>();
+}
